Extract build_libraries_section to modules/output_libraries_section (new module) (#1483)
